fix: #182 충돌 해결

+++ android/data/src/main/java/com/sixkids/data/api/ChallengeService.kt
@@ -20,7 +20,7 @@ interface ChallengeService {
     @POST("challenges")
     suspend fun createChallenge(
         @Body challengeCreateRequest: ChallengeCreateRequest
-    ): ApiResult<ApiResponse<Int>>
+    ): ApiResult<ApiResponse<Long>>
 
     @GET("challenges")
     suspend fun getChallengeHistory(
