+++ android/data/src/main/java/com/sixkids/data/repository/challenge/remote/ChallengeRemoteDataSourceImpl.kt
@@ -26,7 +26,7 @@ class ChallengeRemoteDataSourceImpl @Inject constructor(
         reward: Int,
         minCount: Int,
         groups: List<GroupSimple>,
-    ): Int = challengeService.createChallenge(
+    ): Long = challengeService.createChallenge(
         ChallengeCreateRequest(
             organizationId = organizationId,
             title = title,
